clarify field names

+++ src/rosegold/packets/serverbound/held_item_change.cr
@@ -4,11 +4,11 @@ class Rosegold::Serverbound::HeldItemChange < Rosegold::Serverbound::Packet
   PACKET_ID = 0x25_u32
 
   property \
-    hotbar_nr : UInt16
+    hotbar_nr : UInt8
 
   # `hotbar_nr` ranges from 0 to 8
   def initialize(
-    @hotbar_nr : UInt16
+    @hotbar_nr : UInt8
   ); end
 
   def to_packet : Minecraft::IO
